Move game actions to support folder
